+++ src/main/java/com/ecommercebackoffice/product/entity/Product.java
@@ -25,7 +25,7 @@ public class Product extends BaseEntity {
     private ProductCategory category;
 
     @Column(nullable = false)
-    private Integer price;
+    private Long price;
 
     @Column(nullable = false)
     private Integer stock;
@@ -37,7 +37,7 @@ public class Product extends BaseEntity {
     @Column(nullable = false)
     private Long adminId;
 
-    public Product(String name, ProductCategory category, Integer price,
+    public Product(String name, ProductCategory category, Long price,
                    Integer stock, ProductStatus status, Long adminId) {
         this.name = name;
         this.category = category;
@@ -48,7 +48,7 @@ public class Product extends BaseEntity {
     }
 
     // 상품명, 카테고리, 가격만 수정 가능하도록 제한
-    public void updateInfo(String name, ProductCategory category, Integer price) {
+    public void updateInfo(String name, ProductCategory category, Long price) {
         this.name = name;
         this.category = category;
         this.price = price;
